+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 ### Added
 - module instance support
 
+### Changed
+- registry logger name contains registry type
+
+### Fixed
+- registry configuration change listener handles missing config object gracefully
+
 ## [1.0.0] - 2019-11-27
 ### Added
 - Initial version